Moved fetching of current user id from CardCollection and NewTopicCard component to their respective pages, in order to remove the dependency of modules/topics on modules/platform.
