+++ app/modules/topics/components/CardCollection.js
@@ -4,19 +4,20 @@ import * as React from 'react';
 import { connect } from 'react-redux';
 import { Card } from 'semantic-ui-react';
 
-import { UnsupportedOperationError } from 'errors';
 import type { Identifier } from 'types/model';
 import type { State } from 'types/state';
-import platform from 'modules/platform';
 
 import { getAllTopicIdsByUserId } from '../selectors';
 import { getAllByUserId } from '../actions';
 
 import TopicCard from './TopicCard';
 
+type PassedProps = {|
+  userId: Identifier,
+|};
+
 type StateProps = {|
   topicIds: Array<Identifier>,
-  userId: Identifier,
 |};
 
 type DispatchProps = {|
@@ -24,20 +25,16 @@ type DispatchProps = {|
 |};
 
 type Props = {|
+  ...PassedProps,
   ...StateProps,
   ...DispatchProps,
 |};
 
-const mapStateToProps = (state: State): StateProps => {
-  const userAuth = platform.selectors.getUserAuth(state);
-
-  if (userAuth == null) {
-    throw new UnsupportedOperationError(`This shouldn't happen`);
-  }
+const mapStateToProps = (state: State, props: PassedProps): StateProps => {
+  const { userId } = props;
 
   return {
-    topicIds: getAllTopicIdsByUserId(state, userAuth.userId),
-    userId: userAuth.userId,
+    topicIds: getAllTopicIdsByUserId(state, userId),
   };
 };
