+++ app/pages/LibraryPage.js
@@ -1,10 +1,14 @@
 // @flow
 
 import * as React from 'react';
+import { connect } from 'react-redux';
 import { translate, type TranslatorProps } from 'react-i18next';
 import { Button, Grid } from 'semantic-ui-react';
 import { Link } from 'react-router-dom';
 
+import { type State } from 'types/state';
+import { type Identifier } from 'types/model';
+import { UnsupportedOperationError } from 'errors';
 import Page from 'core-components/Page';
 import platform from 'modules/platform';
 import topics from 'modules/topics';
@@ -12,12 +16,26 @@ import topics from 'modules/topics';
 const { AuthWrapper } = platform.components;
 const { CardCollection } = topics.components;
 
-type Props = TranslatorProps;
+type StateProps = {|
+  currentUserId: Identifier,
+|};
+
+type Props = {| ...TranslatorProps, ...StateProps |};
+
+const mapStateToProps = (state: State): StateProps => {
+  const userAuth = platform.selectors.getUserAuth(state);
+
+  if (userAuth == null) {
+    throw new UnsupportedOperationError(`This shouldn't happen`);
+  }
+
+  return {
+    currentUserId: userAuth.userId,
+  };
+};
 
 const PureLibraryPage = (props: Props): React.Node => {
-  const {
-    t,
-  } = props;
+  const { t, currentUserId } = props;
 
   return (
     <Page>
@@ -37,7 +55,7 @@ const PureLibraryPage = (props: Props): React.Node => {
           </Grid.Row>
           <Grid.Row>
             <Grid.Column width={16}>
-              <CardCollection />
+              <CardCollection userId={currentUserId} />
             </Grid.Column>
           </Grid.Row>
         </Grid>
@@ -46,7 +64,7 @@ const PureLibraryPage = (props: Props): React.Node => {
   );
 };
 
-const LibraryPage = translate()(PureLibraryPage);
+const LibraryPage = connect(mapStateToProps)(translate()(PureLibraryPage));
 
 export { PureLibraryPage };
 export default LibraryPage;
